fix: revert wrong movement of extract slug
