+++ src/utils/invitationLink.ts
@@ -113,33 +113,3 @@ export async function verifyInvitationToken(token: string): Promise<InvitationTo
 
   return decoded;
 }
-
-/**
- * Extract frontend base URL from request headers
- * @param req - Express request object
- * @returns Frontend base URL or null if not found
- */
-export function extractFrontendBaseUrl(req: Request): string | null {
-  // Try to get from Origin header (for CORS requests)
-  if (req.headers.origin) {
-    return req.headers.origin;
-  }
-  // Try to get from Referer header
-  if (req.headers.referer) {
-    // Validate URL format before using
-    if (!req.headers.referer.match(/^https?:\/\/.+/)) {
-      throw new AppException(HttpStatusCode.BadRequest, 'Invalid referer URL format');
-    }
-
-    const url = new URL(req.headers.referer);
-    return `${url.protocol}//${url.host}`;
-  }
-
-  // Try to construct from Host header
-  if (req.headers.host) {
-    const protocol = req.secure || req.headers['x-forwarded-proto'] === 'https' ? 'https' : 'http';
-    return `${protocol}://${req.headers.host}`;
-  }
-
-  return null;
-}
